Add conditional imports to support including the widgets via stubs on VM build too.
A quick example to demonstrate how to conditionally import different versions of the widget that makes it possible to add the seo_renderer also on VM builds. It only returns stubs that don't do anything useful on VM, and only adds the SEO divs on the Web build.

Using a shared abstract class would be nice improvement to ensure they always use the same interface.